% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wpp.R
\name{get_wpp}
\alias{get_wpp}
\title{Download UN DESA WPP data}
\usage{
get_wpp(
  indicator = NULL,
  indicator_file = NULL,
  pop_age = c("total", "single", "five"),
  pop_sex = c("total", "both", "male", "female", "all"),
  pop_freq = c("annual", "five"),
  pop_date = c("jul1", "jan1", "jan1-dec31"),
  variant_id = 2,
  wpp_version = 2022,
  clean_names = FALSE,
  fct_age = TRUE,
  drop_id_cols = FALSE,
  tidy_pop_sex = FALSE,
  add_regions = FALSE,
  add_iso_codes = FALSE,
  messages = TRUE,
  server = c("github", "local")
)
}
\arguments{
\item{indicator}{Character string based on the \code{name} column in the \link[=wpp_indicators]{wpp_indicators} data frame or \code{pop}. Represents the variables to be downloaded.}

\item{indicator_file}{Character string based on the \code{file} column in the \link[=wpp_indicators]{wpp_indicators} data frame . Represents the file group to download data from. Needed for obtaining indicators that are available with different levels of granularities (such as \code{Births} in overall population or \code{Births} by mothers age group).}

\item{pop_age}{Character string for population age groups if \code{indicator} is set to \code{pop}. Defaults to no age groups \code{total}, but can be set to \code{single} or \code{five}.}

\item{pop_sex}{Character string for population sexes if \code{indicator}is set to \code{pop}. Defaults to no sex \code{total}, but can be set to \code{male}, \code{female}, \code{both} or \code{all}.}

\item{pop_freq}{Character string for frequency of population data if \code{indicator} is set to \code{pop}. Defaults to \code{annual}, but in a some (exceptional cases) can be set to \code{five}.}

\item{pop_date}{Character string for frequency of population data if \code{indicator} is set to \code{pop}. Defaults to \code{jul1} (July 1st), but for WPP2022 can be set to \code{jan1} for population at beginning of year or \code{jan1-dec31} for exposure population.}

\item{variant_id}{Numeric value(s) based on the \code{var_id} column in the \link[=wpp_indicators]{wpp_indicators} data frame. Note, past data is in the \code{"Medium" (2)} variant only.}

\item{wpp_version}{Integer for WPP version. Default of \code{2019}. All WPP back to 1998 are available.}

\item{clean_names}{Logical to indicate if column names should be cleaned}

\item{fct_age}{Logical to indicate if \code{AgeGrp} column be converted to a factor.}

\item{drop_id_cols}{Logical to indicate if \code{VarId}, \code{LocID}, \code{MidPeriod}, \code{AgeGrpStart}, \code{AgeGrpSpan} and \code{SexID} columns to be removed.}

\item{tidy_pop_sex}{Logical to indicate if columns for sex specific population data should be stacked into single population column with an accompanying new sex column.}

\item{add_regions}{Logical to indicate if to add a \code{reg_name} and \code{area_name} columns for countries (where \code{LocID} is less than 900)}

\item{add_iso_codes}{Logical to indicate if to add a \code{iso3} and \code{iso2} columns for ISO 3 and 2 letter country codes (where \code{LocID} is less than 900)}

\item{messages}{Logical to not suppress printing of messages.}

\item{server}{Character string for location to download data from. Default of \code{github}.}
}
\value{
A \link[tibble:tibble-package]{tibble} with downloaded data in tidy format
}
\description{
Downloads data on demographic indicators in UN DESA WPP. Requires a working internet connection.
}
\details{
Indicators must use the name corresponding to the \code{name} column in in the \link[=wpp_indicators]{wpp_indicators} data frame.
The \link[=find_indicator]{find_indicator()} function can be used to look up the indicator code and availability by variants

There are 114 different indicators in WPP data (starting from 1998). See the full \href{https://github.com/guyabel/tidywpp#indicator-list}{table} of the different indicators available in each WPP.

The \code{variant_id} argument must be one or more numbers from the \code{var_id} column in the \link[=wpp_indicators]{wpp_indicators} data frame.
Not all indicators area available in all variants. Use the \link[=find_indicator]{find_indicator()} function to check availability.
There are 14 different variants in WPP data (starting from 1998).\tabular{rl}{
   var_id \tab variant \cr
   2 \tab Medium \cr
   3 \tab High \cr
   4 \tab Low \cr
   5 \tab Constant fertility \cr
   6 \tab Instant replacement \cr
   7 \tab Zero migration \cr
   8 \tab Constant mortality \cr
   9 \tab No change \cr
   10 \tab Momentum \cr
   16 \tab Instant replacement zero migration \cr
   202 \tab Median PI (BHM median in WPP2015) \cr
   203 \tab Upper 80 PI \cr
   204 \tab Lower 80 PI \cr
   206 \tab Upper 95 PI \cr
   207 \tab Lower 95 PI \cr
}
}
\examples{
\donttest{
# single indicator from medium variant of latest WPP
get_wpp(indicator = "TFR")

# single indicator from multiple variants of latest WPP
get_wpp(indicator = "TFR", variant_id = c(2, 3, 4))

# some indicators appear in multiple file groups, for example Births
# represents total number of births in the country in the
# Demographic_Indicators file (chosen by default)
get_wpp(indicator = "Births")

# specify indicator_file to get number of births by mothers 5-year age group
get_wpp(indicator = c("Births", "ASFR"),
        indicator_file = "Fertility_by_Age5",
        drop_id_cols = TRUE)

# PopTotal, PopMale and PopFemale indicators are in many WPP files with
# a wide range granularity. Set indicator = "pop" and use the pop_sex,
# pop_age, pop_freq and pop_date to get desired data from the appropriate
# indicator_file...

# when using indicator = "pop" get_wpp() defaults to annual total population
# (summed over age and sex)
get_wpp(indicator = "pop")

# use pop_sex to get specific sexes (or both or all)
get_wpp(indicator = "pop", pop_sex = "male")

# use pop_age to specify age groups
get_wpp(indicator = "pop", pop_sex = "both", pop_age = "five")

# use pop_date to specify populations at start of year (rather than mid-year)
get_wpp(indicator = "pop", pop_sex = "female", pop_age = "five", pop_date = "jan1")

# tidy sex into a single column and drop id columns
get_wpp(indicator = "pop", pop_sex = "both", pop_age = "five",
        tidy_pop_sex = TRUE, drop_id_cols = TRUE)

# alternatively use indicator_file to select the desired version of population indicator(s)
get_wpp(indicator = c("PopTotal", "PopMale", "PopFemale"), indicator_file =  "TotalPopulationBySex")

# clean column names
get_wpp(indicator = c("SRB", "NetMigrations", "PopGrowthRate"),
        clean_names = TRUE, drop_id_cols = TRUE)
}
}
