% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpp_indicators.R
\docType{data}
\name{wpp_indicators}
\alias{wpp_indicators}
\title{WPP indicators}
\format{
A data frame with 2742 rows and 7 variables, including:
\describe{
\item{name}{Short name of indicator. Can be passed to \code{get_wpp()}}
\item{details}{Brief description of indicator.}
\item{var_id}{ID of availabile variants for the indicator}
\item{variant}{Name of availabile variants for the indicator}
\item{file_group}{Group to which the indicator belongs. Indicators in different groups need to downloaded seperatly with \code{get_wpp()}}
\item{file_group_details}{Brief description on the file group}
\item{wpp}{WPP version of the indicator}
}
}
\source{
\url{http://dataexplorer.wittgensteincentre.org/wcde-v2/}
}
\usage{
wpp_indicators
}
\description{
Data set containing the indicator names, details and available variants strating from WPP1998.
}
\keyword{datasets}
