% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wpp.R
\name{get_wpp}
\alias{get_wpp}
\title{Download UN DESA WPP data}
\usage{
get_wpp(
  indicator = NULL,
  indicator_file_group = NULL,
  variant_id = 2,
  wpp_version = 2019,
  clean_names = FALSE,
  fct_age = TRUE,
  drop_id_cols = FALSE,
  tidy_pop_sex = FALSE,
  messages = TRUE
)
}
\arguments{
\item{indicator}{Character string based on the \code{name} column in the \code{wpp_indicators} data frame. Represents the variables to be downloaded.}

\item{indicator_file_group}{Character string based on the \code{file_group} column in the \code{wpp_indicators} data frame . Represents the file group to download data from. Only needed for obtaining different granularities of population data.}

\item{variant_id}{Numeric value(s) based on the \code{var_id} column in the \code{wpp_indicators} data frame. Note, past data is in the \code{"Medium" (2)} variant only.}

\item{wpp_version}{Integer for WPP version. Default of \code{2019}. All WPP back to 1998 are available.}

\item{clean_names}{Logical to indicate if column names should be cleaned}

\item{fct_age}{Logical to indicate if \code{AgeGrp} column be converted to a factor.}

\item{drop_id_cols}{Logical to indicate if \code{VarId}, \code{LocID}, \code{MidPeriod}, \code{AgeGrpStart}, \code{AgeGrpSpan} and \code{SexID} columns to be removed.}

\item{tidy_pop_sex}{Logical to indicate if columns for sex specific population data should be stacked into single population column with an accompanying new sex column.}

\item{messages}{Logical to not suppress printing of messages.}
}
\value{
A tibble with downloaded data in tidy format
}
\description{
Downloads data on demographic indicators in UN DESA WPP. Requires a working internet connection.
}
\details{
Add in table of indicators and variants
}
\examples{
\donttest{
# single indicator from medium variant of latest WPP
get_wpp(indicator = "TFR")

# single indicator from multiple variants of latest WPP
get_wpp(indicator = "TFR", variant_id = c(2, 3, 4))

# multiple population indicators from single variant of
# latest WPP
get_wpp(indicator = c("PopTotal", "PopMale", "PopFemale"))

# as multiple granularities of population in WPP, there are
# multiple population indicators. use indicator_file_group
# to select version of population indicator(s)
get_wpp(indicator = c("PopTotal", "PopMale", "PopFemale"),
        indicator_file_group =  "TotalPopulationBySex")

# tidy sex into a single column and drop id columns
get_wpp(indicator = c("PopMale", "PopFemale"),
        indicator_file_group = "TotalPopulationBySex",
        tidy_pop_sex = TRUE, drop_id_cols = TRUE)

# clean column names
get_wpp(indicator = c("SRB", "NetMigrations", "GrowthRate"),
        clean_names = TRUE, drop_id_cols = TRUE)

# old life table
get_wpp(indicator = c("qx", "lx", "dx", "Lx", "Tx", "ex"),
        wpp_version = 2017, drop_id_cols = TRUE)
}
}
